Feat: 커뮤니티 페이지, 그룹 페이지 구조화
