+++ frontend/src/pages/community/index.vue
@@ -1,24 +1,70 @@
 <template>
   <div>
-    Community
     <h2>My Community</h2>
     <div class="groupBar">
-      <button>그룹 추가</button>
-      <button>그룹 참가</button>
+      <button>
+        <img class="groupIcon" src='@/images/user-group-solid.png'>
+        그룹추가
+      </button>
+      <button>
+        <img id="joinIcon" src="@/images/right-to-bracket-solid.png">
+         그룹참가
+      </button>
     </div>
-    <div>
-      그룹 ...
+    <div class="groupbody">
+      <div class="group1">
+       <a href="url"><img src="@/images/group-img.jpeg" border="0" width="100" height="100"></a>
+       <p>그룹1</p>
+      </div>
+      <div class="group2">
+       <a href="url"><img src="@/images/img.jpg" border="0" width="100" height="100"></a>
+       <p>그룹2</p>
+      </div>
     </div>
   </div>
 </template>
 
 <script>
+
 export default {
   
 }
 </script>
 
 <style scoped>
+  #joinIcon {
+    width: 12px;
+    display: flex;
+    justify-content:center;
+    float: left;
+    padding-left: 4px;
+  }
+  .groupIcon {
+    width: 12px;
+    padding-top: 0px;
+  }
+  .groupbody {
+    text-align: center;
+  }
+  .group1 {
+    text-align: center;
+    border: 1px solid;
+    width: 100px;
+    height: 100px;
+    padding: 50;
+    margin-right: 20px;
+    margin-left: 850px;
+    float: left;
+  }
+  .group2 {
+    text-align: center;
+    border: 1px solid;
+    width: 100px;
+    height: 100px;
+    padding: 50;
+    float: left;
+  }
+  
   .groupBar {
     text-align: center;
   }
@@ -26,7 +72,7 @@ export default {
     text-align: center;
   }
   button {
-    width: 75px;
+    width: 87px;
     height: 25px;
     margin: 5px;
     border-radius: 45px;
