+++ frontend/src/pages/signup/index.vue
@@ -41,7 +41,7 @@ export default {
       userName: "",
       userPhoneNumber: "",
       userBirth: "",
-      // status: true,
+      status: true,
     };
   },
   methods: {
@@ -52,7 +52,7 @@ export default {
         userName: this.userName,
         userPhoneNumber: this.userPhoneNumber,
         userBirth: this.userBirth,
-        // status: this.status,
+        status: this.status,
       };
       const { data } = await registerUser(userData);
       console.log(data)
